Temp

+++ python/install.sh
@@ -1,9 +1,8 @@
 #!/usr/bin/env sh
 
-curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
-python3 get-pip.py --user
-
-rm $(pwd)/get-pip.py
-
-python3 -m pip install --user ansible
-
+if ! pip --version
+then
+    curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
+    python3 get-pip.py --user
+    rm $(pwd)/get-pip.py
+fi
